Detect OS architecture
