+++ foojaysupport/src/main/java/io/foojay/support/FoojayPanel.java
@@ -16,20 +16,16 @@
 package io.foojay.support;
 
 import io.foojay.api.discoclient.pkg.Architecture;
-import io.foojay.api.discoclient.pkg.Bitness;
 import io.foojay.api.discoclient.pkg.Pkg;
 import io.foojay.api.discoclient.pkg.PackageType;
 import io.foojay.api.discoclient.pkg.Distribution;
 import io.foojay.api.discoclient.pkg.OperatingSystem;
-import io.foojay.api.discoclient.pkg.ReleaseStatus;
 import io.foojay.api.discoclient.pkg.VersionNumber;
 import io.foojay.api.discoclient.pkg.ArchiveType;
 import io.foojay.api.discoclient.pkg.Latest;
-import io.foojay.api.discoclient.pkg.LibCType;
 import io.foojay.api.discoclient.pkg.MajorVersion;
-import io.foojay.api.discoclient.pkg.Scope;
-import io.foojay.api.discoclient.pkg.TermOfSupport;
 import io.foojay.api.discoclient.util.Helper;
+import static io.foojay.support.OS.getOperatingSystem;
 import static io.foojay.support.SwingWorker2.submit;
 import java.awt.CardLayout;
 
@@ -43,7 +39,6 @@ import javax.swing.event.ChangeEvent;
 import org.checkerframework.checker.guieffect.qual.UIEffect;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.openide.util.Exceptions;
-import org.openide.util.Utilities;
 
 @SuppressWarnings("initialization")
 public class FoojayPanel extends FirstPanel {
@@ -189,21 +184,6 @@ public class FoojayPanel extends FirstPanel {
             default:
                 throw new IllegalStateException();
         }
-    }
-
-    public static OperatingSystem getOperatingSystem() {
-        if (Utilities.isMac())
-            return OperatingSystem.MACOS;
-        if (Utilities.isWindows())
-            return OperatingSystem.WINDOWS;
-        if (Utilities.isUnix()) {
-            String os = System.getProperty("os.name").toLowerCase();
-            if (os.contains("sunos"))
-                return OperatingSystem.SOLARIS;
-            else
-                return OperatingSystem.LINUX;
-        }
-        return OperatingSystem.NONE;
     }
 
 }
